+++ src/indexer/services/coinmarketcap/coinmarketcap.service.ts
@@ -13,17 +13,15 @@ export type CoinMaketCapResponse = {
 }
 
 class CoinMarketCapService {
-  private coinMarketCapAPIKey: string | undefined
-  private coinMarketCapUrl: string | undefined
+  private coinMarketCapAPIKey: string
+  private coinMarketCapUrl: string
 
-  constructor(coinMakertcapKey?: string, coinMarketcapApiURL?: string) {
+  constructor(coinMakertcapKey: string, coinMarketcapApiURL: string) {
     this.coinMarketCapAPIKey = coinMakertcapKey
     this.coinMarketCapUrl = coinMarketcapApiURL
   }
 
   private async getValueConvertion(amount: string, tokenSymbol: string): Promise<CoinMaketCapResponse["quote"]["USD"]["price"]> {
-    if (!this.coinMarketCapAPIKey || !this.coinMarketCapUrl) throw new Error("CoinMarketCap credentials not found")
-
     const url = `${this.coinMarketCapUrl}/v1/tools/price-conversion?amount=${amount}&symbol=${tokenSymbol}&convert=USD`
 
     try {
